Use symlink for wasmFunctions
